modules/schematics/src/test/java/com/ibm/cloud/schematics/v1/model/BlueprintTest.java
/*
 * (C) Copyright IBM Corp. 2024.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */

package com.ibm.cloud.schematics.v1.model;

import com.ibm.cloud.schematics.v1.model.Blueprint;
import com.ibm.cloud.schematics.v1.model.BlueprintConfigItem;
import com.ibm.cloud.schematics.v1.model.BlueprintFlow;
import com.ibm.cloud.schematics.v1.model.CatalogSource;
import com.ibm.cloud.schematics.v1.model.ExternalSource;
import com.ibm.cloud.schematics.v1.model.GitSource;
import com.ibm.cloud.schematics.v1.model.UserState;
import com.ibm.cloud.schematics.v1.model.VariableData;
import com.ibm.cloud.schematics.v1.model.VariableMetadata;
import com.ibm.cloud.schematics.v1.utils.TestUtilities;
import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
import com.ibm.cloud.sdk.core.util.DateUtils;
import java.io.InputStream;
import java.util.HashMap;
import java.util.List;
import org.testng.annotations.Test;
import static org.testng.Assert.*;

/**
 * Unit test class for the Blueprint model.
 */
public class BlueprintTest {
  final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
  final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();

  @Test
  public void testBlueprint() throws Throwable {
    GitSource gitSourceModel = new GitSource.Builder()
      .computedGitRepoUrl("testString")
      .gitRepoUrl("testString")
      .gitToken("testString")
      .gitRepoFolder("testString")
      .gitRelease("testString")
      .gitBranch("testString")
      .build();
    assertEquals(gitSourceModel.computedGitRepoUrl(), "testString");
    assertEquals(gitSourceModel.gitRepoUrl(), "testString");
    assertEquals(gitSourceModel.gitToken(), "testString");
    assertEquals(gitSourceModel.gitRepoFolder(), "testString");
    assertEquals(gitSourceModel.gitRelease(), "testString");
    assertEquals(gitSourceModel.gitBranch(), "testString");

    CatalogSource catalogSourceModel = new CatalogSource.Builder()
      .catalogName("testString")
      .catalogId("testString")
      .offeringName("testString")
      .offeringVersion("testString")
      .offeringKind("testString")
      .offeringTargetKind("testString")
      .offeringId("testString")
      .offeringVersionId("testString")
      .offeringVersionFlavourName("testString")
      .offeringRepoUrl("testString")
      .offeringProvisionerWorkingDirectory("testString")
      .dryRun(true)
      .owningAccount("testString")
      .itemIconUrl("testString")
      .itemId("testString")
      .itemName("testString")
      .itemReadmeUrl("testString")
      .itemUrl("testString")
      .launchUrl("testString")
      .build();
    assertEquals(catalogSourceModel.catalogName(), "testString");
    assertEquals(catalogSourceModel.catalogId(), "testString");
    assertEquals(catalogSourceModel.offeringName(), "testString");
    assertEquals(catalogSourceModel.offeringVersion(), "testString");
    assertEquals(catalogSourceModel.offeringKind(), "testString");
    assertEquals(catalogSourceModel.offeringTargetKind(), "testString");
    assertEquals(catalogSourceModel.offeringId(), "testString");
    assertEquals(catalogSourceModel.offeringVersionId(), "testString");
    assertEquals(catalogSourceModel.offeringVersionFlavourName(), "testString");
    assertEquals(catalogSourceModel.offeringRepoUrl(), "testString");
    assertEquals(catalogSourceModel.offeringProvisionerWorkingDirectory(), "testString");
    assertEquals(catalogSourceModel.dryRun(), Boolean.valueOf(true));
    assertEquals(catalogSourceModel.owningAccount(), "testString");
    assertEquals(catalogSourceModel.itemIconUrl(), "testString");
    assertEquals(catalogSourceModel.itemId(), "testString");
    assertEquals(catalogSourceModel.itemName(), "testString");
    assertEquals(catalogSourceModel.itemReadmeUrl(), "testString");
    assertEquals(catalogSourceModel.itemUrl(), "testString");
    assertEquals(catalogSourceModel.launchUrl(), "testString");

    ExternalSource externalSourceModel = new ExternalSource.Builder()
      .sourceType("local")
      .git(gitSourceModel)
      .catalog(catalogSourceModel)
      .build();
    assertEquals(externalSourceModel.sourceType(), "local");
    assertEquals(externalSourceModel.git(), gitSourceModel);
    assertEquals(externalSourceModel.catalog(), catalogSourceModel);

    VariableMetadata variableMetadataModel = new VariableMetadata.Builder()
      .type("boolean")
      .aliases(java.util.Arrays.asList("testString"))
      .description("testString")
      .cloudDataType("testString")
      .defaultValue("testString")
      .linkStatus("normal")
      .secure(true)
      .immutable(true)
      .hidden(true)
      .required(true)
      .options(java.util.Arrays.asList("testString"))
      .minValue(Long.valueOf("26"))
      .maxValue(Long.valueOf("26"))
      .minLength(Long.valueOf("26"))
      .maxLength(Long.valueOf("26"))
      .matches("testString")
      .position(Long.valueOf("26"))
      .groupBy("testString")
      .source("testString")
      .build();
    assertEquals(variableMetadataModel.type(), "boolean");
    assertEquals(variableMetadataModel.aliases(), java.util.Arrays.asList("testString"));
    assertEquals(variableMetadataModel.description(), "testString");
    assertEquals(variableMetadataModel.cloudDataType(), "testString");
    assertEquals(variableMetadataModel.defaultValue(), "testString");
    assertEquals(variableMetadataModel.linkStatus(), "normal");
    assertEquals(variableMetadataModel.secure(), Boolean.valueOf(true));
    assertEquals(variableMetadataModel.immutable(), Boolean.valueOf(true));
    assertEquals(variableMetadataModel.hidden(), Boolean.valueOf(true));
    assertEquals(variableMetadataModel.required(), Boolean.valueOf(true));
    assertEquals(variableMetadataModel.options(), java.util.Arrays.asList("testString"));
    assertEquals(variableMetadataModel.minValue(), Long.valueOf("26"));
    assertEquals(variableMetadataModel.maxValue(), Long.valueOf("26"));
    assertEquals(variableMetadataModel.minLength(), Long.valueOf("26"));
    assertEquals(variableMetadataModel.maxLength(), Long.valueOf("26"));
    assertEquals(variableMetadataModel.matches(), "testString");
    assertEquals(variableMetadataModel.position(), Long.valueOf("26"));
    assertEquals(variableMetadataModel.groupBy(), "testString");
    assertEquals(variableMetadataModel.source(), "testString");

    VariableData variableDataModel = new VariableData.Builder()
      .name("testString")
      .value("testString")
      .useDefault(true)
      .metadata(variableMetadataModel)
      .build();
    assertEquals(variableDataModel.name(), "testString");
    assertEquals(variableDataModel.value(), "testString");
    assertEquals(variableDataModel.useDefault(), Boolean.valueOf(true));
    assertEquals(variableDataModel.metadata(), variableMetadataModel);

    BlueprintConfigItem blueprintConfigItemModel = new BlueprintConfigItem.Builder()
      .name("testString")
      .description("testString")
      .source(externalSourceModel)
      .inputs(java.util.Arrays.asList(variableDataModel))
      .build();
    assertEquals(blueprintConfigItemModel.name(), "testString");
    assertEquals(blueprintConfigItemModel.description(), "testString");
    assertEquals(blueprintConfigItemModel.source(), externalSourceModel);
    assertEquals(blueprintConfigItemModel.inputs(), java.util.Arrays.asList(variableDataModel));

    BlueprintFlow blueprintFlowModel = new BlueprintFlow.Builder()
      .build();

    UserState userStateModel = new UserState.Builder()
      .state("draft")
      .setBy("testString")
      .setAt(DateUtils.parseAsDateTime("2019-01-01T12:00:00.000Z"))
      .build();
    assertEquals(userStateModel.state(), "draft");
    assertEquals(userStateModel.setBy(), "testString");
    assertEquals(userStateModel.setAt(), DateUtils.parseAsDateTime("2019-01-01T12:00:00.000Z"));

    Blueprint blueprintModel = new Blueprint.Builder()
      .name("Toronto Dev Environtment")
      .schemaVersion("1.0")
      .source(externalSourceModel)
      .config(java.util.Arrays.asList(blueprintConfigItemModel))
      .description("Deploys dev environtment instance in Toronto Region")
      .resourceGroup("Default")
      .tags(java.util.Arrays.asList("blueprint:Tor-Dev"))
      .location("us-south")
      .inputs(java.util.Arrays.asList(variableDataModel))
      .settings(java.util.Arrays.asList(variableDataModel))
      .flow(blueprintFlowModel)
      .userState(userStateModel)
      .build();
    assertEquals(blueprintModel.name(), "Toronto Dev Environtment");
    assertEquals(blueprintModel.schemaVersion(), "1.0");
    assertEquals(blueprintModel.source(), externalSourceModel);
    assertEquals(blueprintModel.config(), java.util.Arrays.asList(blueprintConfigItemModel));
    assertEquals(blueprintModel.description(), "Deploys dev environtment instance in Toronto Region");
    assertEquals(blueprintModel.resourceGroup(), "Default");
    assertEquals(blueprintModel.tags(), java.util.Arrays.asList("blueprint:Tor-Dev"));
    assertEquals(blueprintModel.location(), "us-south");
    assertEquals(blueprintModel.inputs(), java.util.Arrays.asList(variableDataModel));
    assertEquals(blueprintModel.settings(), java.util.Arrays.asList(variableDataModel));
    assertEquals(blueprintModel.flow(), blueprintFlowModel);
    assertEquals(blueprintModel.userState(), userStateModel);

    String json = TestUtilities.serialize(blueprintModel);

    Blueprint blueprintModelNew = TestUtilities.deserialize(json, Blueprint.class);
    assertTrue(blueprintModelNew instanceof Blueprint);
    assertEquals(blueprintModelNew.name(), "Toronto Dev Environtment");
    assertEquals(blueprintModelNew.schemaVersion(), "1.0");
    assertEquals(blueprintModelNew.source().toString(), externalSourceModel.toString());
    assertEquals(blueprintModelNew.description(), "Deploys dev environtment instance in Toronto Region");
    assertEquals(blueprintModelNew.resourceGroup(), "Default");
    assertEquals(blueprintModelNew.location(), "us-south");
    assertEquals(blueprintModelNew.flow().toString(), blueprintFlowModel.toString());
    assertEquals(blueprintModelNew.userState().toString(), userStateModel.toString());
  }

  @Test(expectedExceptions = IllegalArgumentException.class)
  public void testBlueprintError() throws Throwable {
    new Blueprint.Builder().build();
  }

}